techpack: display: Don't skip LP1 cmd for AoD with fingerprint_mode on
OPlus added extra code to skip sending the LP1/LP2 command if the
fingerprint sensor is active. They probably have other changes they made
in their fork of Android that requires this, but for us, this causes our
panels to not actually enter LP1 mode at all when Ambient Display is
enabled.

Delete special case handling for this so that we can actually use LP1
mode properly.

[Dark-Matter7232: Adjust for our source]
Change-Id: Ia2ae4b8a2e5983036f2128c5681624d83e6b1544
Signed-off-by: Dark-Matter7232 <me@const.eu.org>